Opposite enriched category
I tried to define the opposite of an enriched category.  The data
is easy to get, but, in order to check the axioms, some lemmas on
monoidal and braided categories are required.

These would follow from coherence (which I assume would not be easy
to prove in UniMath), but writing down the complete proof for all of
them would take some time.